requirements.txt
pyyaml
requests
pydantic
openai
anthropic
python-dotenv
psutil
nltk
tabulate
numpy
colorama
ruamel.yaml
nltk